requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
numpy
easyocr
python-dotenv
azure-storage-blob
pandas
ftfy
regex
tqdm